Major tests migration & standarization
Trying to unify code styles in tests and make them run on puppeteer
(installing selenium is a complex thing, and requires to go out of the
project bounds).

There are some tests that I hightly doubt were testing correctly before,
promise results ignored and such.

Special mention to that test that won't pass because the date validation
is better than what the test expected.

However, this rewrite was done with lots of find-and-replace and in a
hurry, so probably there are some tests broken. I may address that
later.